refactor: remove AccountLookupService interface and relocate class
